control plugin registry and storage options via go build tag minimal
